Scaling completely done.
Scaling now visually appealing as well as completely functional. The Scaling Button of the plugin now needs to be removed plus adjustmens for rapidly leaving the plugin need to be done. ALSO: There is an error in the Boolean operations which leads to faulty scale properties - I think. Needs further research for which I dont have time.